python/requirements.txt
langgraph>=1.0.3
langchain-openai>=1.0.2
langgraph-cli[inmem]>=0.4.7
python-dotenv>=1.2.1
typing_extensions>=4.15.0
aiohttp>=3.13.2
a2a-sdk[http-server]>=0.3.0
uvicorn
httpx>=0.27
langchain-core>=0.3.0